requirements.txt
pymongo==4.3.3
python-dotenv==0.21.0
aiogram==2.25.1
